React frontend

+++ src/main/java/ru/payment/calc/payment_calculator/service/NextMonthHoursService.java
@@ -3,8 +3,10 @@ package ru.payment.calc.payment_calculator.service;
 import ru.payment.calc.payment_calculator.model.Group;
 import ru.payment.calc.payment_calculator.model.NextMonthDatesStore;
 
+import java.math.BigDecimal;
+
 public interface NextMonthHoursService {
 
-    double calcNextMonthHours(Group group, NextMonthDatesStore nextMonthDatesStore);
+    BigDecimal calcNextMonthHours(Group group, NextMonthDatesStore nextMonthDatesStore);
 
 }
